moved logic into abstract class for extensions
